feature: log poller for vrf v2/+
