+++ core/services/vrf/v2/listener_v2_types_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/require"
 
-	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/log/mocks"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/vrf_coordinator_v2"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/vrf_coordinator_v2_5"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
@@ -30,7 +29,6 @@ func Test_BatchFulfillments_AddRun(t *testing.T) {
 						TxHash: common.HexToHash("0xd8d7ecc4800d25fa53ce0372f13a416d98907a7ef3d8d3bdd79cf4fe75529c65"),
 					},
 				}),
-				lb: mocks.NewBroadcast(t),
 			},
 			run: pipeline.NewRun(pipeline.Spec{}, pipeline.Vars{}),
 		}, fromAddress)
@@ -49,7 +47,6 @@ func Test_BatchFulfillments_AddRun(t *testing.T) {
 					TxHash: common.HexToHash("0xd8d7ecc4800d25fa53ce0372f13a416d98907a7ef3d8d3bdd79cf4fe75529c65"),
 				},
 			}),
-			lb: mocks.NewBroadcast(t),
 		},
 		run: pipeline.NewRun(pipeline.Spec{}, pipeline.Vars{}),
 	}, fromAddress)
@@ -70,7 +67,6 @@ func Test_BatchFulfillments_AddRun_V2Plus(t *testing.T) {
 						TxHash: common.HexToHash("0xd8d7ecc4800d25fa53ce0372f13a416d98907a7ef3d8d3bdd79cf4fe75529c65"),
 					},
 				}),
-				lb: mocks.NewBroadcast(t),
 			},
 			run: pipeline.NewRun(pipeline.Spec{}, pipeline.Vars{}),
 		}, fromAddress)
@@ -89,7 +85,6 @@ func Test_BatchFulfillments_AddRun_V2Plus(t *testing.T) {
 					TxHash: common.HexToHash("0xd8d7ecc4800d25fa53ce0372f13a416d98907a7ef3d8d3bdd79cf4fe75529c65"),
 				},
 			}),
-			lb: mocks.NewBroadcast(t),
 		},
 		run: pipeline.NewRun(pipeline.Spec{}, pipeline.Vars{}),
 	}, fromAddress)
